+++ lib/donation_system/salesforce/client_error.rb
@@ -1,20 +1,22 @@
 # frozen_string_literal: true
 
-module Salesforce
-  ClientError = Struct.new(:client_error) do
-    def errors
-      error_collection.map { |error| symbolize(error['errorCode']) }
-    end
+module DonationSystem
+  module Salesforce
+    ClientError = Struct.new(:client_error) do
+      def errors
+        error_collection.map { |error| symbolize(error['errorCode']) }
+      end
 
-    private
+      private
 
-    def error_collection
-      return [] unless client_error && client_error.response
-      client_error.response.fetch(:body, [])
-    end
+      def error_collection
+        return [] unless client_error && client_error.response
+        client_error.response.fetch(:body, [])
+      end
 
-    def symbolize(error_code)
-      error_code.downcase.to_sym
+      def symbolize(error_code)
+        error_code.downcase.to_sym
+      end
     end
   end
 end
