+++ lib/donation_system/salesforce/supporter_finder.rb
@@ -1,57 +1,59 @@
 # frozen_string_literal: true
 
-require 'salesforce/client_api'
+require_relative 'client_api'
 
-module Salesforce
-  class SupporterFinder
-    ID = :Id
-    SORT_FIELD = :First_entered__c
+module DonationSystem
+  module Salesforce
+    class SupporterFinder
+      ID = :Id
+      SORT_FIELD = :First_entered__c
 
-    def self.execute(field, value, client = ClientAPI.new)
-      new(client, field, value).execute
-    end
+      def self.execute(field, value, client = ClientAPI.new)
+        new(client, field, value).execute
+      end
 
-    def initialize(client, field, value)
-      @client = client
-      @field = field
-      @value = value
-    end
+      def initialize(client, field, value)
+        @client = client
+        @field = field
+        @value = value
+      end
 
-    def execute
-      Result.new(supporter, errors)
-    end
+      def execute
+        Result.new(supporter, errors)
+      end
 
-    private
+      private
 
-    attr_reader :client, :field, :value
+      attr_reader :client, :field, :value
 
-    def table
-      'Contact'
-    end
+      def table
+        'Contact'
+      end
 
-    def supporter
-      first = first_entered
-      fetch(first[ID]) if first
-    end
+      def supporter
+        first = first_entered
+        fetch(first[ID]) if first
+      end
 
-    def first_entered
-      search.sort_by { |item| item[SORT_FIELD] }.first if search
-    end
+      def first_entered
+        search.sort_by { |item| item[SORT_FIELD] }.first if search
+      end
 
-    def expression
-      "select #{ID}, #{SORT_FIELD} from #{table} where #{field}='#{value}'"
-    end
+      def expression
+        "select #{ID}, #{SORT_FIELD} from #{table} where #{field}='#{value}'"
+      end
 
-    def search
-      @search ||= client.search(expression)
-    end
+      def search
+        @search ||= client.search(expression)
+      end
 
-    def fetch(id)
-      client.fetch(table, id)
-    end
+      def fetch(id)
+        client.fetch(table, id)
+      end
 
-    def errors
-      client.errors
+      def errors
+        client.errors
+      end
     end
   end
 end
