--- lib/donation_system/stripe/result.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Stripe
-  Result = Struct.new(:item, :errors) do
-    def okay?
-      errors.empty?
-    end
-
-    def errors?
-      !okay?
-    end
-  end
-end
