+++ lib/donation_system/thank_you_mailer.rb
@@ -2,13 +2,15 @@
 
 require 'mail'
 
-class ThankYouMailer
-  def self.send_email(recipient, first_name)
-    Mail.deliver do
-      to recipient
-      from 'info@survivalinternational.org'
-      subject 'Thank you for your donation'
-      body "Thank you for your donation #{first_name}"
+module DonationSystem
+  class ThankYouMailer
+    def self.send_email(recipient, first_name)
+      Mail.deliver do
+        to recipient
+        from 'info@survivalinternational.org'
+        subject 'Thank you for your donation'
+        body "Thank you for your donation #{first_name}"
+      end
     end
   end
 end
