+++ spec/donation_system/salesforce/data_structs_for_tests.rb
@@ -1,7 +1,9 @@
 # frozen_string_literal: true
 
-module Salesforce
-  RawSupporterData = Struct.new(:name, :email)
-  RawDonationData = Struct.new(:amount)
-  SupporterSObjectFake = Struct.new(:AccountId)
+module DonationSystem
+  module Salesforce
+    RawSupporterData = Struct.new(:name, :email)
+    RawDonationData = Struct.new(:amount)
+    SupporterSObjectFake = Struct.new(:AccountId)
+  end
 end
