+++ spec/donation_system/salesforce/supporter_creator_spec.rb
@@ -1,50 +1,52 @@
 # frozen_string_literal: true
 
-require 'salesforce/data_structs_for_tests'
-require 'salesforce/supporter_creator'
+require 'donation_system/salesforce/data_structs_for_tests'
+require 'donation_system/salesforce/supporter_creator'
 require 'spec_helper'
 
-module Salesforce
-  RSpec.describe SupporterCreator do
-    let(:data) { RawSupporterData.new('A Name', 'test@test.com') }
+module DonationSystem
+  module Salesforce
+    RSpec.describe SupporterCreator do
+      let(:data) { RawSupporterData.new('A Name', 'test@test.com') }
 
-    describe 'when successful', vcr: { record: :once } do
-      it 'creates a supporter' do
-        result = create_supporter(data)
-        expect(result).to be_okay
-        expect(result.item[:Email]).to eq('test@test.com')
-      end
-    end
-
-    describe 'when unsuccessful', vcr: { record: :once } do
-      it 'fails if there is no data' do
-        result = create_supporter(nil)
-        expect(result).not_to be_okay
-        expect(result.item).to be_nil
+      describe 'when successful', vcr: { record: :once } do
+        it 'creates a supporter' do
+          result = create_supporter(data)
+          expect(result).to be_okay
+          expect(result.item[:Email]).to eq('test@test.com')
+        end
       end
 
-      it 'fails if data is invalid' do
-        result = create_supporter(RawSupporterData.new(nil, nil))
-        expect(result).not_to be_okay
-        expect(result.item).to be_nil
+      describe 'when unsuccessful', vcr: { record: :once } do
+        it 'fails if there is no data' do
+          result = create_supporter(nil)
+          expect(result).not_to be_okay
+          expect(result.item).to be_nil
+        end
+
+        it 'fails if data is invalid' do
+          result = create_supporter(RawSupporterData.new(nil, nil))
+          expect(result).not_to be_okay
+          expect(result.item).to be_nil
+        end
+
+        it 'fails if there is a creation problem' do
+          result = create_supporter(data, ClientFake.new(nil, [:creation_error]))
+          expect(result).not_to be_okay
+          expect(result.item).to be_nil
+        end
       end
 
-      it 'fails if there is a creation problem' do
-        result = create_supporter(data, ClientFake.new(nil, [:creation_error]))
-        expect(result).not_to be_okay
-        expect(result.item).to be_nil
+      def create_supporter(data, client_fake = nil)
+        client_default = ClientAPI.new(Restforce.new(host: 'cs70.salesforce.com'))
+        client = client_fake || client_default
+        described_class.execute(data, client)
       end
-    end
 
-    def create_supporter(data, client_fake = nil)
-      client_default = ClientAPI.new(Restforce.new(host: 'cs70.salesforce.com'))
-      client = client_fake || client_default
-      described_class.execute(data, client)
-    end
-
-    ClientFake = Struct.new(:supporter_id, :errors) do
-      def create(_table, _fields)
-        supporter_id
+      ClientFake = Struct.new(:supporter_id, :errors) do
+        def create(_table, _fields)
+          supporter_id
+        end
       end
     end
   end
